move caller code into its own file
